+++ host-frontend-root/frontend-src-root/src/application/ports/IRewriteRuleRepository.ts
@@ -1,4 +1,4 @@
-import { RewriteRule } from 'src/domain/entities/RewriteRule';
+import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
 
 export interface IRewriteRuleRepository {
   save(rule: RewriteRule): Promise<void>;
