+++ host-frontend-root/frontend-src-root/src/domain/entities/HtmlReplacer.ts
@@ -1,4 +1,4 @@
-import { RewriteRule } from './RewriteRule';
+import { RewriteRule } from './RewriteRule/RewriteRule';
 import { ChildNodeList } from '../value-objects/ChildNodeList';
 import { HtmlContent } from './HtmlContent';
