+++ host-frontend-root/frontend-src-root/src/infrastructure/persistance/storage/ChromeStorageRewriteRuleRepository.ts
@@ -1,5 +1,5 @@
 import { IRewriteRuleRepository } from 'src/application/ports/IRewriteRuleRepository';
-import { RewriteRule } from 'src/domain/entities/RewriteRule';
+import { RewriteRule } from 'src/domain/entities/RewriteRule/RewriteRule';
 
 export class ChromeStorageRewriteRuleRepository implements IRewriteRuleRepository {
   async save(rule: RewriteRule): Promise<void> {
